Move moon phase functions to a separate module
